+++ hutool-core/src/main/java/cn/hutool/core/stream/CollectorUtil.java
@@ -4,6 +4,7 @@ import cn.hutool.core.lang.Opt;
 import cn.hutool.core.util.StrUtil;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.EnumSet;
 import java.util.HashMap;
@@ -246,6 +247,68 @@ public class CollectorUtil {
 					return l;
 				}
 		);
+	}
+
+	/**
+	 * 提供对null值友好的groupingBy操作的{@link Collector}实现，
+	 * 对集合分组，然后对分组后的值集合进行映射
+	 *
+	 * @param classifier       分组依据
+	 * @param valueMapper      值映射方法
+	 * @param valueCollFactory 值集合的工厂方法
+	 * @param mapFactory       Map集合的工厂方法
+	 * @param <T>              元素类型
+	 * @param <K>              键类型
+	 * @param <R>              值类型
+	 * @param <C>              值集合类型
+	 * @param <M>              返回的Map集合类型
+	 * @return {@link Collector}
+	 */
+	public static <T, K, R, C extends Collection<R>, M extends Map<K, C>> Collector<T, ?, M> groupingBy(
+			final Function<? super T, ? extends K> classifier,
+			final Function<? super T, ? extends R> valueMapper,
+			final Supplier<C> valueCollFactory,
+			final Supplier<M> mapFactory) {
+		return groupingBy(classifier, mapFactory, Collectors.mapping(
+				valueMapper, Collectors.toCollection(valueCollFactory)
+		));
+	}
+
+	/**
+	 * 提供对null值友好的groupingBy操作的{@link Collector}实现，
+	 * 对集合分组，然后对分组后的值集合进行映射
+	 *
+	 * @param classifier       分组依据
+	 * @param valueMapper      值映射方法
+	 * @param valueCollFactory 值集合的工厂方法
+	 * @param <T>              元素类型
+	 * @param <K>              键类型
+	 * @param <R>              值类型
+	 * @param <C>              值集合类型
+	 * @return {@link Collector}
+	 */
+	public static <T, K, R, C extends Collection<R>> Collector<T, ?, Map<K, C>> groupingBy(
+			final Function<? super T, ? extends K> classifier,
+			final Function<? super T, ? extends R> valueMapper,
+			final Supplier<C> valueCollFactory) {
+		return groupingBy(classifier, valueMapper, valueCollFactory, HashMap::new);
+	}
+
+	/**
+	 * 提供对null值友好的groupingBy操作的{@link Collector}实现，
+	 * 对集合分组，然后对分组后的值集合进行映射
+	 *
+	 * @param classifier       分组依据
+	 * @param valueMapper      值映射方法
+	 * @param <T>              元素类型
+	 * @param <K>              键类型
+	 * @param <R>              值类型
+	 * @return {@link Collector}
+	 */
+	public static <T, K, R> Collector<T, ?, Map<K, List<R>>> groupingBy(
+			final Function<? super T, ? extends K> classifier,
+			final Function<? super T, ? extends R> valueMapper) {
+		return groupingBy(classifier, valueMapper, ArrayList::new, HashMap::new);
 	}
 
 }