+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * 【core   】     UrlBuilder增加getPortWithDefault方法（pr#835@Gitee）
 * 【core   】     FuncKeyMap的子类，传入可被序列化的keyFunc（pr#838@Gitee）
 * 【extra  】     SpringUtil支持SpringBoot3自动配置（pr#839@Gitee）
+* 【core   】     CollectorUtil添加支持对值集合进行映射的分组方法（pr#844@Gitee）
 
 ### 🐞Bug修复
 * 【poi    】     修复ExcelReader读取只有标题行报错问题（issue#I5U1JA@Gitee）